Replace ArrayList with LinearFifo
LinearFifo is more efficient when regularly removing items from the front of the list.

(Also I wanted to use a FIFO structure when first writing this solution but couldn't find one in the Zig's standard library at the time. Turns out one did exist, it was just called something I didn't think to search for).